scripts/requirements.txt
# requirements.txt for fetch_league_games.py
python-dotenv
requests
